ProtoBuf bug fixed

+++ src/eDoxa.Challenges.Web.Aggregator/Models/BucketModel.cs
@@ -10,6 +10,6 @@ namespace eDoxa.Challenges.Web.Aggregator.Models
     {
         public int Size { get; set; }
 
-        public double Prize { get; set; }
+        public decimal Prize { get; set; }
     }
 }
